requirements.txt
Flask==3.1.3
mysql-connector-python
